+++ stdnet/orm/globals.py
@@ -1,6 +1,6 @@
 import hashlib
 
-from stdnet.utils import to_bytestring, JSPLITTER
+from stdnet.utils import to_bytes, JSPLITTER
 
 __all__ = ['get_model_from_hash', 'hashmodel', 'JSPLITTER']
 
@@ -15,7 +15,7 @@ def hashmodel(model, library = None):
     '''Calculate the Hash id of metaclass ``meta``'''
     library = library or 'python-stdnet'
     meta = model._meta
-    sha = hashlib.sha1(to_bytestring('{0}({1})'.format(library,meta)))
+    sha = hashlib.sha1(to_bytes('{0}({1})'.format(library,meta)))
     hash = sha.hexdigest()[:8]
     meta.hash = hash
     if hash in _model_dict: