+++ stdnet/utils/importer.py
@@ -1,4 +1,4 @@
-try:
+try:    # pragma nocover
     from importlib import *
-except ImportError:
+except ImportError: # pragma nocover
     from .fallbacks._importlib import *