+++ tests/regression/serialize/csv.py
@@ -6,7 +6,7 @@ from examples.data import FinanceTest, Instrument, Fund, Position
 from .base import SerializerMixin
 
 
-class TestFinanceJSON(FinanceTest, SerializerMixin):
+class TestFinanceCSV(FinanceTest, SerializerMixin):
     serializer = 'csv'
     
     def setUp(self):
@@ -18,3 +18,7 @@ class TestFinanceJSON(FinanceTest, SerializerMixin):
         self.assertRaises(ValueError, s.serialize,
                           Fund.objects.query().sort_by('id'))
         self.assertEqual(len(s.data),1)
+        
+    def testLoadError(self):
+        s = self.testDump()
+        self.assertRaises(ValueError, s.load, 'bla')